+++ src/python/fife/extensions/basicapplication.py
@@ -117,6 +117,7 @@ class ApplicationBase:
         engineSetting.setDefaultFontGlyphs(self._finalSetting["FontGlyphs"])
         engineSetting.setDefaultFontPath(self._finalSetting["Font"])
         engineSetting.setDefaultFontSize(self._finalSetting["DefaultFontSize"])
+        engineSetting.setFontPaths(self._finalSetting["FontPaths"])
         engineSetting.setBitsPerPixel(self._finalSetting["BitsPerPixel"])
         engineSetting.setInitialVolume(self._finalSetting["InitialVolume"])
         engineSetting.setSDLRemoveFakeAlpha(self._finalSetting["SDLRemoveFakeAlpha"])